+++ packages/form-utils/src/lib/FormControlLabel.tsx
@@ -1,0 +1,25 @@
+import { forwardRef } from 'react';
+import {
+  Box,
+  FormControlLabel as MuiFormControlLabel,
+  FormControlLabelProps as MuiFormControlLabelProps,
+} from '@mui/material';
+import { FieldHelpIcon } from './FieldHelpIcon';
+
+export type FormControlLabelProps = {
+  /** Adds `FieldHelpIcon` next to the label (should not be within label for accessibility) */
+  helpTopicId?: string;
+  /** If `true`, adds required to control. */
+  required?: boolean;
+} & Omit<MuiFormControlLabelProps, 'labelPlacement' | 'required'>;
+
+export const FormControlLabel = forwardRef(({ helpTopicId, id, ...rest }: FormControlLabelProps, ref) => {
+  const labelId = id || (rest.htmlFor ? `${rest.htmlFor}-label` : undefined);
+
+  return (
+    <Box>
+      <MuiFormControlLabel id={labelId} {...rest} ref={ref} />
+      {helpTopicId ? <FieldHelpIcon helpTopicId="12345" labelId={labelId} /> : null}
+    </Box>
+  );
+});