+++ packages/form-utils/src/lib/FormLabel.stories.tsx
@@ -1,14 +1,12 @@
 // Each exported component in the package should have its own stories file
 
 import type { Meta, StoryObj } from '@storybook/react';
-import { Box, Checkbox, FormControl, InputLabel, OutlinedInput } from '@mui/material';
+import { FormControl, OutlinedInput } from '@mui/material';
 import { FormLabel, FormLabelProps } from './FormLabel';
-import { TextField } from './TextField';
-import { Autocomplete } from './Autocomplete';
-import { FormControlLabel } from './FormControlLabel';
 
+/** Label component for use with individual inputs, form controls,  and form groups. For labels to use with `checkbox`/`radio` see `FormControlLabel`. */
 const meta: Meta<typeof FormLabel> = {
-  title: 'Components/FormLabel/FormLabel',
+  title: 'Components/FormUtils/FormLabel',
   component: FormLabel,
   tags: ['autodocs'],
 };