+++ packages/form-utils/src/lib/FormLabel.tsx
@@ -1,20 +1,20 @@
 import { forwardRef } from 'react';
 import { Box, FormLabel as MuiFormLabel, FormLabelProps as MuiFormLabelProps } from '@mui/material';
-import { FieldHelpIcon } from '@availity/mui-field-help-icon';
+import { FieldHelpIcon } from './FieldHelpIcon';
 
-export interface FormLabelProps extends MuiFormLabelProps {
-  children?: React.ReactNode;
+export type FormLabelProps = {
+  /** Adds `FieldHelpIcon` next to the label (should not be within label for accessibility) */
   helpTopicId?: string;
-}
+} & MuiFormLabelProps;
 
 export const FormLabel = forwardRef<HTMLLabelElement, FormLabelProps>((props, ref) => {
-  const { children, helpTopicId, id, htmlFor, ...rest } = props;
+  const { children, helpTopicId, id, ...rest } = props;
 
-  const labelId = id || (htmlFor ? `${htmlFor}-label` : undefined);
+  const labelId = id || (rest.htmlFor ? `${rest.htmlFor}-label` : undefined);
 
   return (
     <Box display="flex" flexDirection="row">
-      <MuiFormLabel id={labelId} htmlFor={htmlFor} {...rest} ref={ref}>
+      <MuiFormLabel id={labelId} {...rest} ref={ref}>
         {children}
       </MuiFormLabel>
       {helpTopicId ? <FieldHelpIcon helpTopicId={helpTopicId} labelId={labelId} sx={{ px: 0.5 }} /> : null}