+++ packages/element/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.16.0](https://github.com/Availity/element/compare/@availity/element@0.15.0...@availity/element@0.16.0) (2023-08-16)
-
+## [0.17.0](https://github.com/Availity/element/compare/@availity/element@0.16.1...@availity/element@0.17.0) (2023-08-23)
 
 ### Features
 
-* **mui-form-label:** add FormLabel and FieldHelpIcon ([8253874](https://github.com/Availity/element/commit/82538741048aec7a8735016c32e22dcdbea91fc9))
+- **mui-form-utils:** moved fieldhelpicon, formcontrollabel, and formlabel to new form-utils package ([4a68fcb](https://github.com/Availity/element/commit/4a68fcb96c1a826b7f710b30eceff196d62a0da4))
+
+## [0.16.1](https://github.com/Availity/element/compare/@availity/element@0.16.0...@availity/element@0.16.1) (2023-08-23)
+
+## [0.16.0](https://github.com/Availity/element/compare/@availity/element@0.15.0...@availity/element@0.16.0) (2023-08-16)
+
+### Features
+
+- **mui-form-label:** add FormLabel and FieldHelpIcon ([8253874](https://github.com/Availity/element/commit/82538741048aec7a8735016c32e22dcdbea91fc9))
 
 ## [0.15.0](https://github.com/Availity/element/compare/@availity/element@0.14.1...@availity/element@0.15.0) (2023-08-11)