+++ packages/link/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## 0.1.0 (2023-06-05)
+## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-link@0.1.0...@availity/mui-link@0.1.1) (2023-08-23)
 
+## 0.1.0 (2023-06-05)
 
 ### Features
 
-* **mui-link:** add link package ([dbcf569](https://github.com/Availity/element/commit/dbcf569191cd0525b2c474eab0d2b487feb473ee))
+- **mui-link:** add link package ([dbcf569](https://github.com/Availity/element/commit/dbcf569191cd0525b2c474eab0d2b487feb473ee))