+++ packages/theme/CHANGELOG.md
@@ -2,12 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.11.0](https://github.com/Availity/element/compare/@availity/theme@0.10.1...@availity/theme@0.11.0) (2023-08-16)
-
+## [0.12.0](https://github.com/Availity/element/compare/@availity/theme@0.11.2...@availity/theme@0.12.0) (2023-08-23)
 
 ### Features
 
-* **mui-form-label:** add FormLabel and FieldHelpIcon ([8253874](https://github.com/Availity/element/commit/82538741048aec7a8735016c32e22dcdbea91fc9))
+- **mui-form-utils:** moved fieldhelpicon, formcontrollabel, and formlabel to new form-utils package ([4a68fcb](https://github.com/Availity/element/commit/4a68fcb96c1a826b7f710b30eceff196d62a0da4))
+
+## [0.11.2](https://github.com/Availity/element/compare/@availity/theme@0.11.1...@availity/theme@0.11.2) (2023-08-23)
+
+## [0.11.1](https://github.com/Availity/element/compare/@availity/theme@0.11.0...@availity/theme@0.11.1) (2023-08-21)
+
+### Bug Fixes
+
+- **mui-toggle-button:** fix color contrast issues for 508 compliance ([9e6ec3e](https://github.com/Availity/element/commit/9e6ec3e6451efa9638c10a427bba8aac87b3c419))
+
+## [0.11.0](https://github.com/Availity/element/compare/@availity/theme@0.10.1...@availity/theme@0.11.0) (2023-08-16)
+
+### Features
+
+- **mui-form-label:** add FormLabel and FieldHelpIcon ([8253874](https://github.com/Availity/element/commit/82538741048aec7a8735016c32e22dcdbea91fc9))
 
 ## [0.10.1](https://github.com/Availity/element/compare/@availity/theme@0.10.0...@availity/theme@0.10.1) (2023-08-11)